requirements.txt
Flask>=3.0.0
google-cloud-bigquery
gunicorn
Werkzeug>=3.0.0


